separated test cases specific to wagtail 4 onward
