+++ wagtail_streamfield_migration_toolkit/test/apps.py
@@ -2,6 +2,6 @@ from django.apps import AppConfig
 
 
 class WagtailStreamfieldMigrationToolkitTestAppConfig(AppConfig):
-    label = "wagtail_streamfield_migration_toolkit_test"
+    label = "toolkit_test"
     name = "wagtail_streamfield_migration_toolkit.test"
     verbose_name = "Wagtail streamfield-migration-toolkit tests"
