+++ wagtail_streamfield_migration_toolkit/test/factories.py
@@ -1,7 +1,6 @@
 import factory
 from factory.django import DjangoModelFactory
 import wagtail_factories
-from wagtail import VERSION as WAGTAIL_VERSION
 
 from . import models
 
@@ -69,11 +68,3 @@ class SamplePageFactory(wagtail_factories.PageFactory):
 
     class Meta:
         model = models.SamplePage
-
-
-if WAGTAIL_VERSION >= (4, 0, 0):
-    class SampleModelWithRevisionsFactory(DjangoModelFactory):
-        content = wagtail_factories.StreamFieldFactory(BaseStreamBlockFactory)
-
-        class Meta:
-            model = models.SampleModelWithRevisions
